requirements.txt
fpdf2==2.8.2
Pillow==11.1.0
PyMuPDF==1.25.3
PyPDF2==3.0.1
PyQt5==5.15.11
pandas==2.2.3
ttkbootstrap==1.10.1
odfpy==1.4.1
